Add files via upload
Media player that can play songs randomly. It uses the rating of each song to decide the probability of the next song that would be played. After a song has been played, it is removed from the play list. The code is written in c but uses the concepts of Object Oriented design.